+++ tests/settings_backward_compatibility.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 
 use anyhow::{Context, Result};
 
-use pueue_lib::settings::Settings;
+use pueue_lib::settings::{Settings, PUEUE_DEFAULT_GROUP};
 
 /// From 0.15.0 on, we aim to have full backward compatibility.
 /// For this reason, an old (slightly modified) v0.15.0 serialized settings file
@@ -25,7 +25,12 @@ fn test_restore_from_old_state() -> Result<()> {
     let settings: Settings = Settings::read_with_defaults(true, &Some(old_settings_path))
         .context("Failed to read old config with defaults:")?;
 
-    assert!(settings.daemon.groups.get("default").is_some());
+    assert!(settings.daemon.groups.get(PUEUE_DEFAULT_GROUP).is_some());
+    assert_eq!(
+        settings.daemon.groups.get(PUEUE_DEFAULT_GROUP).unwrap(),
+        &1,
+        "The default parallel setting for the 'default' group should be 1."
+    );
 
     Ok(())
 }